Added main activity, routes json file, route classes
The main activity allows to create a new route or select an already
planned one.
The routes.json file is used to store the planned routes. It is still in
development, if you notice anything missing, tell me or modify it.
Location is a simple class for storing a location, as coordinates, as a
name or both. The name may be null or the coordinates absent.
Route contains the info for a route. It contains code to create a new
one from scratch(setters) or a constructor to retrieve it from json.
RoutesContainer is a singlethon class used to get the routes from json
and save them on each change.

If you think something should be modified let me know